Add files via upload
Our team used the MITRE ATT&CK methodology as a thorough guide to analyse the assault, identify the attackers' intrusion tactics, and develop effective response plans in response to the sophisticated cyberattack that targeted a medium-sized financial business. The MITRE ATT&CK paradigm gave us a methodical way to classify and comprehend the many phases of the attack lifecycle, which allowed us to precisely map out the tactics, methods, and procedures (TTPs) of the attacker.
We carefully examined every stage of the assault, from gaining access to removing sensitive material, and compared the observed actions to certain methods described in the MITRE ATT&CK architecture. This procedure made it easier to comprehend how the attackers gained access to the network, elevated their privileges, carried out reconnaissance, and carried out their harmful goals. To improve the organization's security posture and lower the chance of future breaches, we also created customized detection tactics and mitigation measures for each strategy and approach that was found. To increase resistance against comparable cyber dangers, these steps included technical controls, procedural improvements, and personnel training programs.